# contracts/api/v1/disclosure-api.openapi.yaml
openapi: 3.1.0
info:
  title: ga-disclosure — 사람 역할 API(설계사·관리자·준법)
  version: 2.7.0
  description: |
    설계서 §7, 6A 계획 §4. 접두 `/api/v1`은 사람 역할(AGENT·MANAGER·COMPLIANCE)만 닿는다 — 서비스 주체는 404(승인 Q15).
    - 인증: JWT(Bearer) 하나. 역할·설계사·조직은 토큰이 아니라 `identity_link`에서 정한다. 인증 실패는 전부 같은 401(`UNAUTHENTICATED`).
    - 권한 없음·없는 자원(다른 테넌트 포함)·없는 라우트는 같은 404 바이트(`NOT_FOUND`).
    - 오류 본문은 `Problem`(JCS). 업무 거부는 범주로 409(상태 충돌)·422(요청 거부), 본문 `REJECTED` + 거부 코드.
    - 쓰기 POST는 `Idempotency-Key` 필수(없으면 428). 같은 키·같은 요청은 같은 바이트로 재생(`Idempotency-Replayed: true`), 다른 요청 422.
      일회용 자격을 담은 응답(세션 발급)은 재생하지 않는다 — 같은 키의 재요청은 409 `IDEMPOTENCY_NOT_REPLAYABLE`.
    - JSON 응답은 전부 JCS 바이트. 시각은 ISO-8601 UTC(`Instant`), 날짜는 `yyyy-MM-dd`. 스키마 형식 검사는 `format`이 아니라 `pattern`.
    - 목록은 서명된 불투명 커서(`after` → `next`) — 다른 테넌트·다른 목록의 커서는 400 `INVALID_CURSOR`.
servers:
  - url: https://disclosure.example
security:
  - bearerJwt: []
paths:
  /api/v1/disclosures:
    get:
      operationId: listDisclosures
      summary: 확인서 목록(범위로 걸러진 — 설계사 자기 것, 관리자 조직 아래, 준법 테넌트 전체)
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/After'
        - { name: status, in: query, required: false, schema: { $ref: '#/components/schemas/DisclosureStatus' } }
      responses:
        '200':
          description: 목록 한 쪽
          content: { application/json: { schema: { $ref: '#/components/schemas/DisclosurePage' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createDisclosure
      summary: 초안 생성(설계사)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDisclosureRequest' }
            example: { customerRef: CR-0000000000000000000000000000001, groupCode: PG-HEALTH-SIMPLE-NR, consultDate: '2026-09-25', templateType: STANDARD }
      responses:
        '201':
          description: 생성됨(`Location`)
          headers:
            Location: { schema: { type: string } }
          content: { application/json: { schema: { $ref: '#/components/schemas/DisclosureReceipt' } } }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/disclosures/{disclosureId}/items:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: replaceItems
      summary: 항목 전체 교체(산출 이후면 비교 전으로 되돌린다)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemsRequest' }
            example:
              items:
                - { productKey: 'INS-A:PRD-1001', recommended: true }
                - { insurerCode: INS-D, productName: '(가상) 임시 상품', quoteDocNo: Q-2026-0001 }
      responses:
        '200': { $ref: '#/components/responses/DisclosureReceipt' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/compare:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: compare
      summary: 비교 완료(단계 검증)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { $ref: '#/components/responses/DisclosureReceipt' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/grades:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: requestGrades
      summary: 엔진 등급·순위 산출(재산출 포함)과 정합성 검증
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { $ref: '#/components/responses/DisclosureReceipt' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/recommendations:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: setRecommendations
      summary: 추천사유(설계사 입력만)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecommendationsRequest' }
            example: { reasons: [ { itemNo: 1, codes: [PREMIUM] }, { itemNo: 3, codes: [COVERAGE], text: '보장 범위 우선' } ] }
      responses:
        '200': { $ref: '#/components/responses/DisclosureReceipt' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/validate:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: validate
      summary: 검증 미리보기(상태를 바꾸지 않는다)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ValidateRequest' }
            example: { stage: SEAL }
      responses:
        '200':
          description: 규칙별 결과
          content: { application/json: { schema: { $ref: '#/components/schemas/ValidationReceipt' } } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/seal:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: seal
      summary: 봉인·발급
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 봉인됨
          content: { application/json: { schema: { $ref: '#/components/schemas/SealReceipt' } } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/void:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: voidDisclosure
      summary: 무효(사유 코드)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LifecycleRequest' }
            example: { reasonCode: DUPLICATE }
      responses:
        '200': { $ref: '#/components/responses/LifecycleReceipt' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/abandon:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: abandonDraft
      summary: 초안 폐기(봉인 전, 고정 룰의 사유 코드 draft.abandonReasons — 텍스트 없음). 추천사유·항목 입력값·청약번호를 지운 묘비가 된다
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AbandonRequest' }
            example: { reasonCode: CUSTOMER_DECLINED }
      responses:
        '200': { $ref: '#/components/responses/LifecycleReceipt' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/rebase:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: rebase
      summary: 재기준(룰 교체로 막힌 초안)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { $ref: '#/components/responses/LifecycleReceipt' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/exception-approvals:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    post:
      operationId: approveException
      summary: 예외 승인(관리자) — 실패 대상 해시에 귀속
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExceptionApprovalRequest' }
            example: { ruleId: R-TEMP-PRODUCT, subjectHash: 0f1e2d3c4b5a69788796a5b4c3d2e1f00f1e2d3c4b5a69788796a5b4c3d2e1f0, reason: '발행번호 확인' }
      responses:
        '200':
          description: 승인 기록
          content: { application/json: { schema: { $ref: '#/components/schemas/ExceptionApprovalReceipt' } } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/artifacts/{kind}:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
      - { name: kind, in: path, required: true, schema: { type: string, enum: [CANONICAL_JSON, PDF, SIGNED_PDF, EVIDENCE_ZIP] } }
    get:
      operationId: getArtifact
      summary: 봉인 산출물 평문 바이트(열람 감사). 산출물이 없거나 읽을 수 없으면 409.
      responses:
        '200':
          description: 산출물 바이트 그대로
          content:
            application/pdf: { schema: { type: string, contentMediaType: application/pdf } }
            application/json: { schema: { type: object } }
            application/zip: { schema: { type: string, contentMediaType: application/zip } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/disclosures/{disclosureId}/anchor-receipt:
    parameters:
      - $ref: '#/components/parameters/DisclosureId'
    get:
      operationId: getAnchorReceipt
      summary: 앵커 영수증 내보내기(준법) — `anchor-receipt-export` 스키마 바이트 그대로. 아직 덮이지 않았으면 409.
      responses:
        '200':
          description: 영수증 바이트
          content: { application/json: { schema: { type: object } } }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 발급자·대상·서명·만료 검증. 클레임은 `sub`·`tenant_id`만 읽는다.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, pattern: '^[A-Za-z0-9_-]{16,128}$' }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    After:
      name: after
      in: query
      required: false
      description: 앞 쪽 응답의 `next`(서명된 불투명 커서)
      schema: { $ref: '#/components/schemas/Cursor' }
    DisclosureId:
      name: disclosureId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
  responses:
    Problem:
      description: 오류(인증 401·없음 404·형식 400·키 없음 428·업무 거부 409/422·작업 겹침 409 등)
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
    DisclosureReceipt:
      description: 적용됨
      content: { application/json: { schema: { $ref: '#/components/schemas/DisclosureReceipt' } } }
    LifecycleReceipt:
      description: 적용됨
      content: { application/json: { schema: { $ref: '#/components/schemas/LifecycleReceipt' } } }
...
  schemas:
    Uuid: { type: string, pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$' }
    Instant: { type: string, pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}(:\d{2}(\.\d{1,9})?)?Z$' }
    Date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
    Sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
    Code: { type: string, pattern: '^[A-Z][A-Z0-9_]{0,63}$' }
...
    Cursor: { type: string, pattern: '^[A-Za-z0-9_-]+\.[A-Za-z0-9_-]+$', maxLength: 512 }
    DisclosureNo: { type: string, pattern: '^[A-Z0-9][A-Z0-9_]{0,31}-\d{4}-\d{6}$' }
...
    NullableInstant:
      oneOf:
        - { type: 'null' }
        - { $ref: '#/components/schemas/Instant' }
    DisclosureStatus:
      type: string
      enum: [DRAFT, COMPARED, GRADED, REASONED, SEALED, PARTIALLY_SIGNED, COMPLETED, VOID, SUPERSEDED, EXPIRED, ABANDONED]
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          enum: [UNAUTHENTICATED, NOT_FOUND, METHOD_NOT_ALLOWED, MALFORMED_REQUEST, REJECTED, CONFLICT, CONCURRENT_WRITE, JOB_ALREADY_RUNNING,
                 REPORT_NOT_AVAILABLE, IDEMPOTENCY_KEY_REQUIRED, IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS, IDEMPOTENCY_NOT_REPLAYABLE,
                 INVALID_CURSOR, RATE_LIMITED, INTERNAL_ERROR]
        message: { type: string, description: 코드별 고정 영문 문장(입력값·자원 값 없음) }
        details:
          oneOf:
            - { type: object, additionalProperties: false, maxProperties: 0 }
            - type: object
              additionalProperties: false
              required: [field]
              properties:
                field: { type: string, pattern: '^[A-Za-z][A-Za-z0-9-]{0,63}$' }
            - type: object
              additionalProperties: false
              required: [rejections]
              properties:
                rejections:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    additionalProperties: false
                    required: [code]
                    properties:
                      code: { $ref: '#/components/schemas/Code' }
                      ruleId: { type: string, pattern: '^R-[A-Z0-9-]{1,63}$' }
...
    DisclosureSummary:
      type: object
      additionalProperties: false
      required: [disclosureId, disclosureNo, version, status, agentId, customerRef, groupCode, consultDate, sealedAt, destroyedAt]
      properties:
        disclosureId: { $ref: '#/components/schemas/Uuid' }
        disclosureNo: { oneOf: [ { type: 'null' }, { $ref: '#/components/schemas/DisclosureNo' } ] }
        version: { type: integer, minimum: 1 }
        status: { $ref: '#/components/schemas/DisclosureStatus' }
        agentId: { type: string }
        customerRef: { type: string, description: 고객 가명 참조(개인정보 아님) }
        groupCode: { type: string }
        consultDate: { $ref: '#/components/schemas/Date' }
        sealedAt: { $ref: '#/components/schemas/NullableInstant' }
        destroyedAt: { $ref: '#/components/schemas/NullableInstant' }
    DisclosurePage:
      type: object
      additionalProperties: false
      required: [items, next]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/DisclosureSummary' } }
        next: { oneOf: [ { type: 'null' }, { $ref: '#/components/schemas/Cursor' } ] }
...
    CreateDisclosureRequest:
      type: object
      additionalProperties: false
      required: [customerRef, groupCode, consultDate, templateType]
      properties:
        customerRef: { type: string, minLength: 1 }
        groupCode: { type: string, minLength: 1 }
        consultDate: { $ref: '#/components/schemas/Date' }
        templateType: { type: string, enum: [STANDARD, AUTO] }
        applicationNo: { type: string, pattern: '^\S{1,64}$', description: '(6B) 선택 청약번호 — 계약 연결의 첫 매칭 키, 작성 뒤 변경 불가' }
    ItemsRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: false
            description: 카탈로그 상품은 productKey, 임시등록은 insurerCode·productName·quoteDocNo. 플래그는 생략하면 false.
            properties:
              productKey: { type: string, pattern: '^[A-Z0-9_-]+:[A-Za-z0-9_.-]+$' }
              insurerCode: { type: string }
              productName: { type: string }
              quoteDocNo: { type: string }
              recommended: { type: boolean }
              requestedByCustomer: { type: boolean }
              agentValues: { type: object }
    RecommendationsRequest:
      type: object
      additionalProperties: false
      required: [reasons]
      properties:
        reasons:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [itemNo, codes]
            properties:
              itemNo: { type: integer, minimum: 1 }
              codes: { type: array, items: { $ref: '#/components/schemas/Code' } }
              text: { type: string }
    ValidateRequest:
      type: object
      additionalProperties: false
      required: [stage]
      properties:
        stage: { type: string, enum: [COMPARE, GRADE, REASON, SEAL] }
    LifecycleRequest:
      type: object
      additionalProperties: false
      required: [reasonCode]
      properties:
        reasonCode: { $ref: '#/components/schemas/Code' }
        reasonText: { type: string }
    AbandonRequest:
      type: object
      additionalProperties: false
      required: [reasonCode]
      properties:
        reasonCode: { $ref: '#/components/schemas/Code' }
    ExceptionApprovalRequest:
      type: object
      additionalProperties: false
      required: [ruleId, subjectHash, reason]
      properties:
        ruleId: { type: string }
        subjectHash: { $ref: '#/components/schemas/Sha256' }
        reason: { type: string, minLength: 1 }
    DisclosureReceipt:
      type: object
      additionalProperties: false
      required: [disclosureId, status]
      properties:
        disclosureId: { $ref: '#/components/schemas/Uuid' }
        status: { $ref: '#/components/schemas/DisclosureStatus' }
    ValidationReceipt:
      type: object
      additionalProperties: false
      required: [disclosureId, results]
      properties:
        disclosureId: { $ref: '#/components/schemas/Uuid' }
        results:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [ruleId, passed, overridable, subjectHash]
            properties:
              ruleId: { type: string }
              passed: { type: boolean }
              overridable: { type: boolean }
              subjectHash: { oneOf: [ { type: 'null' }, { $ref: '#/components/schemas/Sha256' } ] }
    SealReceipt:
      type: object
      additionalProperties: false
      required: [disclosureId, status, disclosureNo, retentionPending]
      properties:
        disclosureId: { $ref: '#/components/schemas/Uuid' }
        status: { $ref: '#/components/schemas/DisclosureStatus' }
        disclosureNo: { oneOf: [ { type: 'null' }, { $ref: '#/components/schemas/DisclosureNo' } ] }
        retentionPending: { type: boolean }
    LifecycleReceipt:
      type: object
      additionalProperties: false
      required: [disclosureId, status, newVersionId]
      properties:
        disclosureId: { $ref: '#/components/schemas/Uuid' }
        status: { $ref: '#/components/schemas/DisclosureStatus' }
        newVersionId: { oneOf: [ { type: 'null' }, { $ref: '#/components/schemas/Uuid' } ] }
    ExceptionApprovalReceipt:
      type: object
      additionalProperties: false
      required: [reviewId, disclosureId, ruleId]
      properties:
        reviewId: { $ref: '#/components/schemas/Uuid' }
        disclosureId: { $ref: '#/components/schemas/Uuid' }
        ruleId: { type: string }
...